+++ packages/perl_xml_namespacesupport.rb
@@ -1,0 +1,24 @@
+require 'package'
+
+class Perl_xml_namespacesupport < Package
+  description 'XML::NamespaceSupport - A simple generic namespace processor'
+  homepage 'https://metacpan.org/pod/XML::NamespaceSupport'
+  version '1.12'
+  license 'GPL-1+ or Artistic'
+  compatibility 'all'
+  source_url 'https://cpan.metacpan.org/authors/id/P/PE/PERIGRIN/XML-NamespaceSupport-1.12.tar.gz'
+  source_sha256 '47e995859f8dd0413aa3f22d350c4a62da652e854267aa0586ae544ae2bae5ef'
+
+  def self.prebuild
+    system 'perl', 'Makefile.PL'
+    system "sed -i 's,/usr/local,#{CREW_PREFIX},g' Makefile"
+  end
+
+  def self.build
+    system 'make'
+  end
+
+  def self.install
+    system 'make', "DESTDIR=#{CREW_DEST_DIR}", 'install'
+  end
+end
